AAA-163: implement isEnabled flag in user model
Make it possible to utilize "isEnabled" for users.  This has been
broken since inception since the folks submitted the code back in
2014, and was recently caught.  Basically, it was not implemented
period in the user lookup logic.

Additionally, changes were needed to the "User" API-- Jaxb
serialization was never working.  That is, the original authors
decided to use an "int" to represent enabled in H2, yet, tried to
map to a Boolean in the serdes process.  This means that calls to
User.isEnabled() always returned null, which is VERY bad practice.

The public methods were left in place, some wrappers were added, and
the internal implementation was changed to utilize an int for
User.enabled field, but still provide boolean the User.isEnabled()
predicate.

Change-Id: I92aa981035d91ca6ee1836d00546b446c7dc9738
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>