Eliminate infinite busy wait in KarafIniWebEnvironment
We've seen SFT's hang on shut down due to the busy wait if the
blueprint container for aaa-shiro fails or is destroyed while
being created. The busy wait blocks startup of the web app and also
blocks on shut down causing the hang. To alleviate this, use a
Future to obtain the AAAShiroProvider. Also added a BundleActivator
to register a blueprint event listener that fails the Future
exceptionally when the blueprint container is destroyed.

Change-Id: I9198579aa026ab685af73ee624c353d7097ba86e
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>