add new API for programmatic registration of web Servlet, Filter, etc.
implementation(s) for this can be seen in the next commit ("chained")

The purpose of this API is to let projects with web components, such as
neutron, aaa, restconf, ditch their respective web.xml.  This will have
a number of advantages, some of which are documented in the JavaDoc of
the new WebServer interface and WebContext class.

see also discussion and interest from project neutron re. adoption on:
https://lists.opendaylight.org/pipermail/neutron-dev/2018-February/001587.html

This is the change originally raised in infrautils as
Ib2df87ca31a2bde547efbf73e0475a1cd64ea6ea proposed to aaa for political
reasons, as discussed during the Kernel Projects call on 2018/02/27.

Change-Id: Ib2fb02aa19e49aa482062f18ba84124a9a623364
Signed-off-by: Michael Vorburger <vorburger@redhat.com>