AAA-168: Remove embedded h2 dependency
This was code inherited from a long time ago and punted around AAA
without any real cleanup.  I am not sure why the original authors
decided to embed the dependency rather than just import it, but this
causes several issues.  This patch removes the embedded h2 dependency
in favor of direct import.  While I recognize that other parts of
ImportPackage need to be cleaned, they will be done in a subsequent
patch since they are separate concepts than what is done here.
In other words, expect follow-ups to continuing cleaning.

aaa-cli-jar relied on aaa-shiro shading the com.h2database:h2 jar,
so I instead added it as a compile time dependency for the module
and extracted the appropriate files for the generated jar in the
maven-shade-plugin configuration.

Change-Id: I9267f1373ddc5b8af0304fd5719dcc96b8874c32
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>