Update patch set 1
Patch Set 1:

As far as I can tell from a manual quick AAA test, it still works.  But I have not tested this very extensively, and will let others -1 this if they find it breaks anything in AAA.  I *DO NOT* want to mix this in with c/68840, so that if it breaks something, this could be reverted without that.

Patch-set: 1