make PaxWebServer ServiceFactory fail instead of return bogus WebServer
The current implementation may return a bogus defunct noop WebServer
which just ignores Servlet & Filter registrations (and just logs a WARN,
which could easily be overlooked) in case it cannot find the Pax Web
WebContainer service.

This change makes it instead "fail fast" on the WebServer service look
up.  Note that the ServiceFactory doc explicitly allows throwing
exceptions; that will be caught and turn into a null service reference,
which whatever is trying to obtain the WebServer must gracefully handle.

Change-Id: Ie4fcbdb125095d353d466958fade98fc759aefd4
Signed-off-by: Michael Vorburger <vorburger@redhat.com>