Update patch set 3
Patch Set 3: Code-Review+1

(4 comments)

> Why didn't you just -1 my patch?

because I wanted that to go in ASAP for all this to finally move (which it did), and see this as a minor improvement on top of that.  Isn't it more productive to work like this together and constructively propose improvementss instead of the silly -1 blocking of each other's work?  I like how this moves forward now!

 > Patch Set 2: Code-Review-1

What was this 2nd -1 here for?  For me this one could go in now...

Patch-set: 3
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1