Update patch set 3
Patch Set 3:

(3 comments)

This looks good. The other option would be to add the requireAuthentication method to the WebContextBuilder to avoid the need for a separate service. Internally PaxWebServer can delegate to the shiro authentication service. This would avoid a potential mistake if someone built the WebContext before calling the Securer.

Patch-set: 3