Update patch set 3
Patch Set 3:

> > (3 comments)
 > >
 > > This looks good. The other option would be to add the
 > > requireAuthentication method to the WebContextBuilder to avoid
 > the
 > > need for a separate service. Internally PaxWebServer can delegate
 > > to the shiro authentication service. This would avoid a potential
 > > mistake if someone built the WebContext before calling the
 > Securer.
 > 
 > The only thing is that cause a circular service dependency between
 > aaa-shiro and aaa-web-impl. Have to think about that....  perhaps
 > split out the servlet registration stuff into a separate aaa-shiro
 > bundle.

Or just call the ShiroWebContextSecurer directly from the shiro WebInitializer.

So basically what I'm saying is to do what you've done here but add a convenience requireAuthentication method on WebContextBuilder to make it easier for normal users so they don't have to import 2 services and coordinate them.

Patch-set: 3