Update patch set 1
Patch Set 1:

> Maybe it's fine - I see similar in other generated features. Just
 > wondering. I assume we don't want error-prone shipped at runtime?

error-prone itself is just another build time quality control too.  What you picked up upon above is not error-prone (the tool) itself though, but it's annotations JAR.  I'm not 100% sure if it's OK to artifically remove that.. there is that whole thing about annotations having a scope, but then in theory even if they are not needed at runtime but exist in the class the JVM ignoring them, but then under OSGi they end up in Import-Package so you need them after all, unless oyu make them optional Maven dependencies... I remember battling with this when I first worked on getting FindBugs into ODL.  Today it does not seem to be worth the effort to avoid a JAR that is a few KBs.  If you think this should be further investigated, let's have a JIRA about it (for project ODLPARENT not for AAA, I guess), but not hold up this change?

Patch-set: 1
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0