Update patch set 4
Patch Set 4:

> 
 > No, that's not a good idea IMHO, and mixes up two unrelated
 > concerns which could well be implemented by completely separate
 > service API implementations.  One of my key goals here is to
 > eventually completely remove any dependency to AAA Shiro from
 > projects using the new Web API.  This will open the door both to
 > more independently evolve aaa-shiro internals, and allows for
 > possible alternative implementations, later. For example, in the
 > sort of web tests I would like to write in Neutron, I want to be
 > able plug a NOOP WebContextSecurer.  You don't want to have to
 > subclass and modify the WebServer implementation for that, because
 > "registering Servlet and Filters on a Web Server" (implemented by
 > Jetty or another web server, directly or through OSGi Karaf Pax
 > Web) is really quite a separate and orthogonal concern from "do
 > shit to protect a web context" (implemented by AAA, Shiro or some
 > other mechanism, or a NOOP not even in AAA).

Maybe they're technically separate in terms of APIs - you use one API to register the servlets etc and another API to secure. But that's an implementation detail. That doesn't mean we can't combine them in the higher-level WebServer API. To me that makes sense b/c no one will want to deploy a web app w/o authentication in production - so why not make it simpler and ore natural for users. There can still be a WebContextSecurer API exactly as you've done that's used internal by the PaxWebServer impl. Also don't you think it make more sense and is cleaner for someone to do:

  WebContext.builder()...requireAuthentication(...).build;

rather than:

  WebContextBuilder webContextBuilder = ....
  webContextSecurer.requireAuthentication(webContextBuilder, ...);
  webContextBuilder.build();

Re: tests, you can just set a NOOP WebServer impl in the WebServer impl.

That said, I'd like to see others weigh in on this (we can discuss on the kernel call) -
 if the consensus is to do it the way you've suggested then that's fine.

Patch-set: 4