Remove static ServiceProxy
ServiceProxy exists to allow the aaa-shiro-act bundle to activate the
AAAFilter using statics. In lieu of statics, I added a AAAShiroActivation
interface implemented by AAAFilter advertised as an OSGi service that
is consumed by aaa-shiro-act to activate the AAAFilter directly.

I also renamed the AAAShiroProvider class in aaa-shiro-act to
AAAShiroActivator for more clarity and so it's not duplicated with
the AAAShiroProvider in aaa-shiro.

Change-Id: I655dba8c42a981315aa543560e10ea9181d87a34
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>