DO NOT MERGE: AAA -> Jersey2
This is an initial stab at converting AAA to Jersey 2 in an
isolated manner.  For this test, ancillary classes such as
some Realms and tests were deleted because they aren't needed
for POC purposes.  This patch allows odl-aaa-shiro to come up
utilizing Jersey 2.

However, this causes a package uses constraint violation for
javax.ws.rs-api in RESTCONF.  It is provided here (for Jersey 2)
as well as by Jersey 1.17 in a repackaged manner.

This patch is for Jersey 2.25;  Jersey 2.26 is newest but cannot
be used yet since it relies on the newer javax.ws.rs-api 2.1.0.

Change-Id: Ie63eb22bfee6557126713f79c71945d5b39ab05d
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>