WIP: Convert AAA to use web/servlet-api
This is an initial stab at converting AAA to Jersey 2 in an isolated
manner. For this test, ancillary classes such as some Realms and tests
were deleted because they aren't needed for POC purposes. This patch
allows odl-aaa-shiro to come up utilizing Jersey 2.

However, this causes a package uses constraint violation for
javax.ws.rs-api in RESTCONF. It is provided here (for Jersey 2) as well
as by Jersey 1.17 in a repackaged manner.

Rework activation works in concert with web/servlet-api, so that the
application is completely oblivious what implmentation is used to
create the servlet.

Change-Id: Ie63eb22bfee6557126713f79c71945d5b39ab05d
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>