DO NOT MERGE:  Shiro-backed AAAEncryptionService impl
We should not be writing our own code for anything Encryption related.
There is a significant amount of custom code inside
AAAEncryptionServiceImpl that is not maintainable.  Additionally, Shiro
provides crypto libraries that are much easier than the JCA-provided
ones.

This change does two things.
1) Introduces a shiro-backed AAAEncryptionService.  An adaptor is
created to bridge between the existing API (lackluster) and the new
implementation.  This implementation makes use of a shared private
AES encryption key.  In this iteration (Not Mergable!) the key is
created on instantiation;  in the future the key will be created
and stored in persistent storage of some kind.

2) Sets the adaptor as the default implementation.  Users can still
use the old version by selecting odl:type="original" in blueprint.
This allows data created by existing ODL implementations to still
work, even though ideally the data will be migrated instead.

Migration is another concern.  For now, we aren't offering a migration
utility, since most known users of the encryption service are done
in a context where the data won't be maintained over an upgrade
(i.e., passwords for NETCONF mount point).  The solution might be to
set up defaults to mirror the existing impl.

Change-Id: I62a5863e0acae4a564292d0a8e52af8092fcad05
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>