Update patch set 4
Patch Set 4: Code-Review+1

> Build Started https://jenkins.opendaylight.org/releng/job/aaa-patch-test-genius-fluorine/2/

There was a weird hiccup in the first genius gate job I ran, but it was ok
on a re-run. the netvirt 1node gate was fine, and the 3node job fell all over
itself in ways that we are used to it falling all over itself.

So, +1 from me, FWIW.

Would this be a cherry pick to oxygen candidate? I think things are a tad bit
more stable there and I would like to be able to investigate this new "car
purchase" failure there if possible.

Patch-set: 4
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1