Allow custom Filters advertised as OSGi services
Currently, custom filters are hooked in via the
org.opendaylight.aaa.filterchain.cfg file. The Filter
instances are instantiated and initialized via reflection.
This mechanism is useful for pre-existing filters to be
dynamically registered externally. However it's a bit
cumbersome for components that want to hook in their
own Filter in code like the TSDR restconf collector. It
has to manipulate the .cfg file and inject dependencies
statically. To alleviate this, support was added to allow
a custom Filter to be advertised as a service.

Change-Id: Id5572386bf65e06bfaf19e6463d6d45e9412fb5b
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>