make JettyWebServer accept WebContext path without slash (and test it)
because that is, apparently, what the "real" (Jetty through Pax Web)
implementation in the Karaf run-time does.  This has come up in the
review of change I614a48bcc425db8faa1a79ebd6236236bcd09e73, where I
added a slash to a context to get a test to pass - with this, that is no
longer needed.

Change-Id: I3c632adbb4c4a9d9ac7bed4557bcaaece47dd0cf
Signed-off-by: Michael Vorburger <vorburger@redhat.com>