Convert web-osgi-impl to use OSGi DS
Declarative Services are a much nicer fit than blueprint for
the use case at handle. SCR can deal with most of the inter-bundle
complexity by having a component with scope=BUNDLE -- lowering
the amount of actual code we need.

Also mark all of this for future evolution, as the OSGi R7 certainly
changes things a bit.

Change-Id: I74e1a7eb34e1b1e94ec751588df0c2c643ce8efc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>