Add PasswordCredentialAuth
CredentialAuth is a generic interface which needs to be specialized
to capture capture which credentials are being passed. Downstreams
(in netconf) have a FIXME related to this, as they have no way to
express the type-safe dependency.

Introduce PasswordCredentialAuth to which downstreams can safely
bind. Also add a method to allow run-time discovery of required
credetial type.

Change-Id: I844db3b460bf95110fb3adbb687ce25e996e3608
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>