Convert aaa-password-service to OSGi DS
This patch serves as a blueprint of what a replacement for blueprint's
clustered-app-config needs to look like. Since we are replicating
the functionality of config DS -> component, this ends up being more
complicated than we'd want it.

Providing the infrastructure replacement for odl:clustered-config
is part of MDSAL-598. This patch provides one possibility how that would
be realized.

JIRA: AAA-203
JIRA: MDSAL-598
Change-Id: I760a6514ed185fc60639f7d7195159d19bf60a76
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>