Migrate aaa-encrypt-service to OSGi DS
This service is using clustered-app-config, hence the migration
is a bit more involved, but this gets rid of another blueprint.

The way this operates is it instantiates listens for datastore changes,
updates the configuration if needed, and then uses a ComponentFactory
to instantiate the service.

One notable change here is that we no longer provide an .xml which the
user can edit -- hence the datastore is only populated if it is empty
and we generate a new encryption key for every new deployment.

JIRA: AAA-204
Change-Id: I83a8f8fea8e272dc4b9c801be6799a15aa71f5ec
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>
Signed-off-by: OleksandrZharov <Oleksandr.Zharov@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>