Throw exception if decryption/encryption fails
Previously, in the case of a failure during encryption or decryption
in AAA, the system would only log an error and return insered string.
This could lead user to believe that the operation was successful,
resulting in them receiving unencrypted/undecrypted data.
Also is possible that IllegalArgumentException is thrown
which is also wrong.

Now, is throw an custom DecryptionException/EncryptionException
to properly handle failures.

JIRA: AAA-266
Change-Id: I4c9078e293fe5b98f0e6b69568ca10a75a4fbe07
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>