Create Filter for jetty Request
The Authentication field is set to UNAUTHENTICATED because
the PaxWebServletHandler could not set this field due to
the missing remote.user attribute.

To resolve this create custom aaa filter
JettyAuthenticationLogFilter.
This filter sets the Authentication field with the user from
Basic authorization to ensure the correct name is printed
in the NCSA logs.

To enable this filter is required to activate it in Karaf console:
scr:enable org.opendaylight.aaa.filter.JettyAuthenticationLogFilter

JIRA: AAA-282
Change-Id: I26c1975ab30f0e57f0c57519d4cc1cc0becdc826
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>