Validate WebContext.name()
WebContext.name() is used to derive HTTP_WHITEBOARD_CONTEXT_NAME, which
is required to conform to symbolic-name syntax.

Users are guided by 'descriptive name' part of the contract and are
using bad names, which has reprocussions all over the place.

This patch adds explicit validation of name, so that we reject such
attempts.

Change-Id: I520cf1d3d9a998bec56610dced7a3ffe128b5e95
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>